Introduce Llama guard customization notebook and associated dataset loader example
This  notebook explores the customization of Llama Guard 3 for specific application needs. Llama Guard, a versatile AI safety tool, can be adapted to maximise its relevance in various scenarios.

We start with zero-shot prompting, a powerful method that allows Llama Guard to make predictions without prior explicit examples. This technique is particularly useful for initial explorations and quick setups. As we progress, we'll delve into adding and removing safety categories before touching on evaluation and fine-tuning processes, where we adjust Llama Guard's parameters to better align with our specific data and use cases. By following the steps in this notebook, the reader should gain a solid understanding of how to tailor Llama Guard 3 effectively, ensuring it performs optimally for your unique requirements. The notebook does not cover every aspect of Llama Guard 3, but focusses on vaious aspects of customization.